Final test for token operations
