+++ gatekeeper-cli/src/commands/fetch-token-id.ts
@@ -46,7 +46,7 @@ import {
       const gatewayNetwork = await makeGatewayNetworkTs(parsedFlags)
       const gatewayTokenTs = await makeGatewayTs(parsedFlags)
 
-      const networkId = await gatewayNetwork.getNetworkId(utils.formatBytes32String(args.networkName));
+      const networkId = await gatewayNetwork.getNetworkId(args.networkName);
       const tokenId = await gatewayTokenTs.getTokenId(tokenOwner, networkId.valueOf() as bigint, args.onlyActive);
